changes added for parameter apis

+++ src/main/java/com/backoffice/operations/service/impl/CmsControlParameterServiceImpl.java
@@ -1,6 +1,8 @@
 package com.backoffice.operations.service.impl;
 
 import com.backoffice.operations.entity.CmsControlParameter;
+import com.backoffice.operations.entity.TransferParameter;
+import com.backoffice.operations.exceptions.DuplicateEntryException;
 import com.backoffice.operations.payloads.CmsControlParameterDTO;
 import java.util.List;
 import java.util.Optional;
@@ -8,13 +10,17 @@ import java.util.stream.Collectors;
 
 import com.backoffice.operations.repository.CmsControlParameterRepository;
 import com.backoffice.operations.service.CmsControlParameterService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class CmsControlParameterServiceImpl implements CmsControlParameterService {
 
     private final CmsControlParameterRepository repository;
+    private static final Logger LOGGER = LoggerFactory.getLogger(CmsControlParameterServiceImpl.class);
 
     @Autowired
     public CmsControlParameterServiceImpl(CmsControlParameterRepository repository) {
@@ -37,8 +43,16 @@ public class CmsControlParameterServiceImpl implements CmsControlParameterServic
 
     @Override
     public CmsControlParameterDTO createParameter(CmsControlParameterDTO parameterDTO) {
-        CmsControlParameter parameter = mapDTOToEntity(parameterDTO);
-        return mapEntityToDTO(repository.save(parameter));
+        try {
+            CmsControlParameter parameter = mapDTOToEntity(parameterDTO);
+            return mapEntityToDTO(repository.save(parameter));
+        }catch (DataIntegrityViolationException e){
+            LOGGER.error("Error occurred while saving cms control parameter", e);
+            throw new DuplicateEntryException("Duplicate entry");
+        }catch (Exception e) {
+            LOGGER.error("Error occurred while saving cms control parameter", e);
+            throw new RuntimeException("Something went wrong");
+        }
     }
 
     @Override
@@ -54,8 +68,20 @@ public class CmsControlParameterServiceImpl implements CmsControlParameterServic
     }
 
     @Override
-    public void deleteParameter(String id) {
-        repository.deleteById(id);
+    public boolean deleteParameter(String id) {
+        try {
+            Optional<CmsControlParameter> existingParameterOptional = repository.findById(id);
+            if (existingParameterOptional.isPresent()) {
+                repository.deleteById(id);
+                return true;
+            } else {
+                LOGGER.warn("No cms control parameter found with ID: {}", id);
+                return false;
+            }
+        } catch (Exception e) {
+            LOGGER.error("Error occurred while deleting cms control parameter by ID: {}", id, e);
+            return false;
+        }
     }
 
     private CmsControlParameterDTO mapEntityToDTO(CmsControlParameter parameter) {
